chore(performance): add users for legacy correspondences (#1339)
* Added users with legacy data to legacy test

* add log output and staging(tt02)

* check if configmap exist before trying to create it

* use valid serviceowner-attributes

* removed debug

* use randomItem

* new data files

+++ Test/Altinn.Correspondence.LoadTests/correspondence/run-test-in-k8s.sh
@@ -135,16 +135,18 @@ if [ ! -f "archive.tar" ]; then
 fi
 
 # Create the configmap from the archive
-if ! kubectl create configmap $configmapname --from-file=archive.tar; then
-    echo "Error: Failed to create configmap"
-    rm archive.tar
-    exit 1
+if ! kubectl get configmap $configmapname &>/dev/null; then
+  if ! kubectl create configmap $configmapname --from-file=archive.tar; then
+      echo "Error: Failed to create configmap"
+      rm archive.tar
+      exit 1
+  fi
 fi
 
 # Create the config.yml file from a string
-arguments="--out experimental-prometheus-rw --vus=$vus --duration=$duration --tag testid=$testid --log-output=none"
+arguments="--out experimental-prometheus-rw --vus=$vus --duration=$duration --tag testid=$testid"
 if $breakpoint; then
-    arguments="--out experimental-prometheus-rw --tag testid=$testid --log-output=none"
+    arguments="--out experimental-prometheus-rw --tag testid=$testid"
 fi
 
 # Create the config.yml file from a string
